+++ TestAssemblies/TestAssembly/AbstractClassWithMethod.cs
@@ -1,0 +1,7 @@
+﻿namespace TestAssembly
+{
+    public abstract class AbstractClassWithMethod
+    {
+        public abstract void Foo();
+    }
+}