+++ TestAssemblies/TestAssembly/IInterfaceWithProperty.cs
@@ -1,0 +1,7 @@
+﻿namespace TestAssembly
+{
+    public interface IInterfaceWithProperty
+    {
+        string Foo { get; }
+    }
+}