+++ TestAssemblies/TestAssembly/IMarkerInterfaceWithInheritance.cs
@@ -1,0 +1,6 @@
+﻿namespace TestAssembly
+{
+    public interface IMarkerInterfaceWithInheritance : IMarkerInterface
+    {
+    }
+}