+++ TestAssemblies/TestAssembly/NonAbstractClass.cs
@@ -1,0 +1,9 @@
+﻿namespace TestAssembly
+{
+    public class NonAbstractClass : IInterfaceWithMethod
+    {
+        public void Bar()
+        {
+        }
+    }
+}